feat: Add Divide Two Integers

+++ Math/Divide Two Integers/java/Solution.java
@@ -1,6 +1,6 @@
 class Solution {
     public int divide(int dividend, int divisor) {
-        int a=dividend/divisor;
+        double a=dividend/divisor;
         return a;
     }
 }
